[IDSDEV-996] and [IDSDEV-1008] Fix Review Changes

+++ components/gsma-authenticators/src/main/java/com/wso2telco/gsma/authenticators/DBUtils.java
@@ -401,6 +401,7 @@ public class DBUtils {
                 authenticationData.setStatus(rs.getInt("status"));
 
             }
+
         } catch (SQLException ex) {
             log.error("authenticationData Error " + ex);
         } finally {
@@ -411,7 +412,6 @@ public class DBUtils {
                     log.error("Error " + e);
                 }
             }
-
             return authenticationData;
         }
     }
@@ -1123,15 +1123,27 @@ public class DBUtils {
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
         StringBuilder sql = new StringBuilder();
-        sql.append("DELETE from consent_given_user_lifetime ");
+        sql.append("SELECT msisdn from consent_given_user_lifetime ");
         sql.append("where msisdn =?");
         try {
             connection = getConnectDBConnection();
             preparedStatement = connection.prepareStatement(sql.toString());
             preparedStatement.setString(1, msisdn);
-            preparedStatement.execute();
+            resultSet = preparedStatement.executeQuery();
+            preparedStatement.close();
+            if(resultSet.next()){
+                sql = new StringBuilder();
+                sql.append("DELETE from consent_given_user_lifetime ");
+                sql.append("where msisdn =?");
+                preparedStatement = connection.prepareStatement(sql.toString());
+                preparedStatement.setString(1, msisdn);
+                preparedStatement.execute();
+            } else{
+                log.info("No records found to delete ");
+            }
         }catch (SQLException e) {
-            log.info("No records found to delete ");
+            log.error("Error in removing Approved APIs for MIG User " + e.getMessage());
+            throw new AuthenticatorException();
         }  finally {
             IdentityDatabaseUtil.closeAllConnections(connection, resultSet,preparedStatement);
         }
@@ -1165,7 +1177,7 @@ public class DBUtils {
                 scope_role.add(resultSet.getString("scope_role"));
             }
         }catch (SQLException e) {
-            log.info("No roles found to the scope ");
+            handleException("Error while getting roles mapped with scopes ", e);
         }  finally {
             IdentityDatabaseUtil.closeAllConnections(connection, resultSet,preparedStatement);
         }
